+++ src/modules/read-file.cpp
@@ -59,18 +59,18 @@ void initializeImageArray(ifstream *file, int file_type, uint32_t image_number, 
 
 void printFiles(uint32_t number_of_images, uint32_t number_of_query_images, uint64_t d) {
     cout << "INPUT DATASET:" << endl;
-    for (int i = 0; i < number_of_images; i++){
+    for (uint32_t i = 0; i < number_of_images; i++){
         cout << "next image" << endl;
-        for (int j = 0; j < d; j++) {
+        for (uint64_t j = 0; j < d; j++) {
             cout << static_cast<unsigned>(all_images[i][j]) << '\t';
         }
         cout << endl;
     }
     cout << "QUERY DATASET:" << endl;
-    for (int i = 0; i < number_of_query_images; i++)
+    for (uint32_t i = 0; i < number_of_query_images; i++)
     {
         cout << "next image" << endl;
-        for (int j = 0; j < d; j++) {
+        for (uint64_t j = 0; j < d; j++) {
             cout << static_cast<unsigned>(query_images[i][j]) << '\t';
         }
         cout << endl;
